backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
motor==3.3.2
pydantic==2.5.3
pymongo==4.6.1
python-dotenv==1.0.0
email-validator==2.1.0
python-multipart==0.0.6 